Refactor Julc compiler's PirGenerator for improved readability and maintainability
