+++ julc-compiler/src/test/resources/golden/foreach-single-acc.flat.hex
@@ -1,0 +1,1 @@
+010100232533357340022930b191919b8948000004c8cc004dd61aba1357446ae88d5d11aba2357446ae88d5d11aba235573c6ea800920003232300223300200200123002233002002001222533357346ae8c00840044cc00cd5d10011900099b8000148008d5d09aab9e3754003
