+++ julc-compiler/src/test/resources/golden/foreach-single-acc.uplc
@@ -1,0 +1,1 @@
+(program 1.1.0 (lam scriptContextData [(lam result__ (force [[[(force (builtin ifThenElse)) i1] (delay (con unit ()))] (delay (error))])) [(lam txInfo [(lam count [[(builtin lessThanEqualsInteger) (con integer 0)] i1]) [(lam loop__forEach__0 [[i1 [(builtin unListData) [(force (builtin headList)) [(force (builtin tailList)) [(force (builtin tailList)) [(force (builtin tailList)) [(force (builtin tailList)) [(force (builtin tailList)) [(force (builtin tailList)) [(force (builtin tailList)) [(force (builtin tailList)) [(force (force (builtin sndPair))) [(builtin unConstrData) i2]]]]]]]]]]]]] (con integer 0)]) [(lam f [(lam x [i2 (lam v [[i2 i2] i1])]) (lam x [i2 (lam v [[i2 i2] i1])])]) (lam loop__forEach__0 (lam xs__ (lam count (force [[[(force (builtin ifThenElse)) [(force (builtin nullList)) i2]] (delay i1)] (delay [[i3 [(force (builtin tailList)) i2]] [(lam count i1) [[(builtin addInteger) i1] (con integer 1)]]])]))))]]]) [(force (builtin headList)) [(force (force (builtin sndPair))) [(builtin unConstrData) i1]]]]]))
