+++ julc-compiler/src/test/resources/golden/hof-filter.flat.hex
@@ -1,0 +1,1 @@
+0101002323232533357340022930b191919b87323300148000008c8c8c0088cc0080080048c0088cc0080080048894ccd5cd1aba300110021330033370000490011aba200148008c8cc004dd8a4c64660026ec5260033232300223300200200123002233002002001222533357346ae8c00440084cc00ccc894ccd5cd19b8848010dd6800899aba00010021002002357420026ae88004c8c8c0088cc0080080048c0088cc0080080048894ccd5cd1aba30011002133003323357400020066ae84004d5d10009bac001357426ae88004d55cf1baa0011
