+++ julc-compiler/src/test/resources/golden/hof-filter.uplc
@@ -1,0 +1,1 @@
+(program 1.1.0 (lam scriptContextData [(lam ctxFields__ [(lam redeemer__ [(lam result__ (force [[[(force (builtin ifThenElse)) i1] (delay (con unit ()))] (delay (error))])) [(lam items [(lam filtered [[(builtin equalsInteger) [(lam go__f [[i1 (con integer 0)] i2]) [(lam f [(lam x [i2 (lam v [[i2 i2] i1])]) (lam x [i2 (lam v [[i2 i2] i1])])]) (lam go__f (lam acc__f (lam lst__f (force [[[(force (builtin ifThenElse)) [(force (builtin nullList)) i1]] (delay i2)] (delay [[i3 [[(builtin addInteger) i2] (con integer 1)]] [(force (builtin tailList)) i1]])]))))]]] (con integer 1)]) [(lam go [[i1 [(builtin mkNilData) (con unit ())]] [(lam go [[i1 [(builtin mkNilData) (con unit ())]] i3]) [(lam f [(lam x [i2 (lam v [[i2 i2] i1])]) (lam x [i2 (lam v [[i2 i2] i1])])]) (lam go (lam acc (lam lst (force [[[(force (builtin ifThenElse)) [(force (builtin nullList)) i1]] (delay i2)] (delay [[i3 [[(lam acc_flt (lam x_flt (force [[[(force (builtin ifThenElse)) [[(builtin lessThanInteger) (con integer 2)] [(builtin unIData) i1]]] (delay [[(force (builtin mkCons)) i1] i2])] (delay i2)]))) i2] [(force (builtin headList)) i1]]] [(force (builtin tailList)) i1]])]))))]]]) [(lam f [(lam x [i2 (lam v [[i2 i2] i1])]) (lam x [i2 (lam v [[i2 i2] i1])])]) (lam go (lam acc (lam lst (force [[[(force (builtin ifThenElse)) [(force (builtin nullList)) i1]] (delay i2)] (delay [[i3 [(lam x_rev [[(force (builtin mkCons)) i1] i3]) [(force (builtin headList)) i1]]] [(force (builtin tailList)) i1]])]))))]]]) [(builtin unListData) i1]]]) [(force (builtin headList)) [(force (builtin tailList)) i1]]]) [(force (force (builtin sndPair))) [(builtin unConstrData) i1]]]))
