+++ julc-compiler/src/test/resources/golden/hof-map.flat.hex
@@ -1,0 +1,1 @@
+0101002323232533357340022930b191919b87375a6ae840052004323300137629319198009bb149800cc8c8c0088cc0080080048c0088cc0080080048894ccd5cd1aba30011002133003323357406ea0cdc01bad0014800800cd5d08009aba20013232300223300200200123002233002002001222533357346ae8c00440084cc00cc8cd5d00008019aba1001357440026eb0004d5d09aba200135573c6ea8005
