+++ julc-compiler/src/test/resources/golden/hof-map.uplc
@@ -1,0 +1,1 @@
+(program 1.1.0 (lam scriptContextData [(lam ctxFields__ [(lam redeemer__ [(lam result__ (force [[[(force (builtin ifThenElse)) i1] (delay (con unit ()))] (delay (error))])) [(lam items [(lam mapped [[(builtin equalsInteger) [(builtin unIData) [(force (builtin headList)) i1]]] (con integer 2)]) [(lam go [[i1 [(builtin mkNilData) (con unit ())]] [(lam go [[i1 [(builtin mkNilData) (con unit ())]] i3]) [(lam f [(lam x [i2 (lam v [[i2 i2] i1])]) (lam x [i2 (lam v [[i2 i2] i1])])]) (lam go (lam acc (lam lst (force [[[(force (builtin ifThenElse)) [(force (builtin nullList)) i1]] (delay i2)] (delay [[i3 [(lam x_map [[(force (builtin mkCons)) [(builtin iData) [[(builtin addInteger) [(builtin unIData) i1]] (con integer 1)]]] i3]) [(force (builtin headList)) i1]]] [(force (builtin tailList)) i1]])]))))]]]) [(lam f [(lam x [i2 (lam v [[i2 i2] i1])]) (lam x [i2 (lam v [[i2 i2] i1])])]) (lam go (lam acc (lam lst (force [[[(force (builtin ifThenElse)) [(force (builtin nullList)) i1]] (delay i2)] (delay [[i3 [(lam x_rev [[(force (builtin mkCons)) i1] i3]) [(force (builtin headList)) i1]]] [(force (builtin tailList)) i1]])]))))]]]) [(builtin unListData) i1]]]) [(force (builtin headList)) [(force (builtin tailList)) i1]]]) [(force (force (builtin sndPair))) [(builtin unConstrData) i1]]]))
