+++ julc-compiler/src/test/resources/golden/multi-acc-while.flat.hex
@@ -1,0 +1,1 @@
+010100232533357340022930b19191919b873370000400290211bad357426ae88008dd69aba10013230013357406ea120003357406ea120003357406ea120063762931919180111980100100091801119801001000912999ab9a323371090000009bad357426ae88d5d1000898011919191919191919aba0375000666ae80dd400119aba037500026ec52633702006900119b8000348050cdc0001a40046eb4d5d09aba2357440066eb4d5d09aba2002375a6ae8400400440041
