+++ julc-compiler/src/test/resources/golden/nested-while-no-acc.flat.hex
@@ -1,0 +1,1 @@
+010100232533357340022930b191919b8700148030dd69aba10013230013357406ea120003357406ea120063762931919180111980100100091801119801001000912999ab9a323371090000009bad357426ae880044c008c8c8c8c8c8cd5d01ba80023357406ea0004dd8a4c66e040092002337000040026eb4d5d09aba2002375a6ae8400400440041
