+++ julc-compiler/src/test/resources/golden/nested-while-no-acc.uplc
@@ -1,0 +1,1 @@
+(program 1.1.0 (lam scriptContextData [(lam result__ (force [[[(force (builtin ifThenElse)) i1] (delay (con unit ()))] (delay (error))])) [(lam __t [(lam sum [[(builtin equalsInteger) i1] (con integer 6)]) [(builtin unIData) [(force (builtin headList)) i1]]]) [(lam loop__while__1 [i1 [[(force (builtin mkCons)) [(builtin iData) (con integer 0)]] [[(force (builtin mkCons)) [(builtin iData) (con integer 3)]] [(builtin mkNilData) (con unit ())]]]]) [(lam f [(lam x [i2 (lam v [[i2 i2] i1])]) (lam x [i2 (lam v [[i2 i2] i1])])]) (lam loop__while__1 (lam __acc_tuple (force [[[(force (builtin ifThenElse)) [(lam i [[(builtin lessThanInteger) (con integer 0)] i1]) [(builtin unIData) [(force (builtin headList)) [(force (builtin tailList)) i1]]]]] (delay [i2 [(lam __t [(lam sum [(lam i [(lam sum [(lam i [[(force (builtin mkCons)) [(builtin iData) i2]] [[(force (builtin mkCons)) [(builtin iData) i1]] [(builtin mkNilData) (con unit ())]]]) [[(builtin subtractInteger) i2] (con integer 1)]]) [[(builtin addInteger) i2] i1]]) [(builtin unIData) [(force (builtin headList)) [(force (builtin tailList)) i2]]]]) [(builtin unIData) [(force (builtin headList)) i1]]]) i1]])] (delay i1)])))]]]]))
