+++ julc-compiler/src/test/resources/golden/nested-while.flat.hex
@@ -1,0 +1,1 @@
+010100232533357340022930b191919b870014831804dd69aba10013230013357406ea120003357406ea120003762931919180111980100100091801119801001000912999ab9a323371000290031bad357426ae880044c008c8c8c8c8c8c8cd5d01ba80023357406ea0004dd8a4c66e0000d2002375a6ae84004c8c004cd5d01ba80033357406ea120003762931919180111980100100091801119801001000912999ab9a323371000290031bad357426ae880044c008c8c8c8c8c8cd5d01ba80023357406ea0004dd8a4c66e0000920023370066e00008cdc1003240280026eb4d5d09aba2002375a6ae840040044004dd69aba1357440046eb4d5d0800800880081
