+++ julc-compiler/src/test/resources/golden/simple-validator.flat.hex
@@ -1,0 +1,1 @@
+0101002323232533357340022930b1919b870014832004cdc00008019aba1357440026aae78dd50009
