+++ julc-compiler/src/test/resources/golden/simple-validator.uplc
@@ -1,0 +1,1 @@
+(program 1.1.0 (lam scriptContextData [(lam ctxFields__ [(lam redeemer__ [(lam result__ (force [[[(force (builtin ifThenElse)) i1] (delay (con unit ()))] (delay (error))])) [(lam sum [[(builtin equalsInteger) i1] (con integer 100)]) [[(builtin addInteger) i1] i3]]]) [(force (builtin headList)) [(force (builtin tailList)) i1]]]) [(force (force (builtin sndPair))) [(builtin unConstrData) i1]]]))
