+++ julc-compiler/src/test/resources/golden/while-break.uplc
@@ -1,0 +1,1 @@
+(program 1.1.0 (lam scriptContextData [(lam result__ (force [[[(force (builtin ifThenElse)) i1] (delay (con unit ()))] (delay (error))])) [(lam __t [(lam acc [[(builtin lessThanInteger) (con integer 0)] i1]) [(builtin unIData) [(force (builtin headList)) i1]]]) [(lam loop__while__0 [i1 [[(force (builtin mkCons)) [(builtin iData) (con integer 0)]] [[(force (builtin mkCons)) [(builtin iData) (con integer 5)]] [(builtin mkNilData) (con unit ())]]]]) [(lam f [(lam x [i2 (lam v [[i2 i2] i1])]) (lam x [i2 (lam v [[i2 i2] i1])])]) (lam loop__while__0 (lam __acc_tuple (force [[[(force (builtin ifThenElse)) [(lam counter [[(builtin lessThanInteger) (con integer 0)] i1]) [(builtin unIData) [(force (builtin headList)) [(force (builtin tailList)) i1]]]]] (delay [(lam __t [(lam acc [(lam counter (force [[[(force (builtin ifThenElse)) [[(builtin lessThanInteger) (con integer 3)] i2]] (delay [[(force (builtin mkCons)) [(builtin iData) i2]] [[(force (builtin mkCons)) [(builtin iData) i1]] [(builtin mkNilData) (con unit ())]]])] (delay [(lam acc [(lam counter [i7 [[(force (builtin mkCons)) [(builtin iData) i2]] [[(force (builtin mkCons)) [(builtin iData) i1]] [(builtin mkNilData) (con unit ())]]]]) [[(builtin subtractInteger) i2] (con integer 1)]]) [[(builtin addInteger) i2] i1]])])) [(builtin unIData) [(force (builtin headList)) [(force (builtin tailList)) i2]]]]) [(builtin unIData) [(force (builtin headList)) i1]]]) i1])] (delay i1)])))]]]]))
